Move template routes to api/config.ts
